// apps/realtime-server/src/canvas/contracts/canvas-types.ts
export type CanvasRoomRef = {
  canvasId: string;
  workspaceId: string;
};

export type CanvasPresenceUser = {
  color?: string;
  displayName?: string;
  userId: string;
};

export type CanvasPresencePoint = {
  x: number;
  y: number;
};

export type CanvasPresenceViewport = {
  height: number;
  width: number;
  x: number;
  y: number;
  zoom: number;
};

export type CanvasLoadedViewportBounds = {
  height: number;
  margin: number;
  width: number;
  x: number;
  y: number;
};

export type CanvasRoomLoadedRegion = {
  bottom: number;
  id: string;
  left: number;
  loadedAt: string;
  right: number;
  top: number;
};